+++ frontend/app/src/components/history/events/HistoryEventsList.vue
@@ -106,7 +106,7 @@ const [DefineTable, ReuseTable] = createReusableTemplate();
           <div
             v-for="(item, index) in events"
             :key="index"
-            class="grid md:grid-cols-8 py-4 items-center"
+            class="grid md:grid-cols-4 gap-x-2 gap-y-4 lg:grid-cols-9 xl:grid-cols-10 py-4 items-center"
             :class="{
               'border-b border-default': index < events.length - 1
             }"
@@ -114,7 +114,7 @@ const [DefineTable, ReuseTable] = createReusableTemplate();
             <HistoryEventType
               :event="item"
               :chain="getChain(item.location)"
-              class="md:col-span-2"
+              class="md:col-span-2 lg:col-span-3"
             />
             <HistoryEventAsset :event="item" class="md:col-span-2" />
             <HistoryEventNote
@@ -123,7 +123,7 @@ const [DefineTable, ReuseTable] = createReusableTemplate();
               :chain="getChain(item.location)"
               :no-tx-hash="isNoTxHash(item)"
               :block-number="blockEvent?.blockNumber"
-              class="break-words leading-6 md:col-span-3"
+              class="break-words leading-6 md:col-span-3 xl:col-span-4"
             />
             <RowActions
               class="md:col-span-1"